u8 -> u32

+++ Rust/to_hex.rs
@@ -1,4 +1,4 @@
-fn to_hex (chr: u8) -> Vec<String> {
+fn to_hex (chr: u32) -> Vec<String> {
     let mut hex_operand = chr;
     let mut vec = Vec::new();
 
@@ -31,7 +31,7 @@ fn to_hex (chr: u8) -> Vec<String> {
     vec
 }
 
-fn alphabet_match (hex: u8) -> String {
+fn alphabet_match (hex: u32) -> String {
     let alphabet;
     
     match hex {
@@ -48,8 +48,14 @@ fn alphabet_match (hex: u8) -> String {
 }
 
 fn main() {
+    // let a = "abcde";
+    // let mut vec = Vec::new();
     let a = 63;
-    let vec = to_hex(a as u8);
+    let vec = to_hex(a as u32);
+    // for c in a.chars() {
+    //     print!("{} ", to_hex(c as u32));
+    //     vec.push(c as u32);
+    // }
     print!("{} hex -> ", a);
     for v in vec {
         print!("{}", v);
